efficient catalog typing
